The Beginning - First Commit
Everything has to start somewhere, and this project is no different.
This is a test for something bigger, and at this point will be at its
messiest. But should clean up as things move forward. Got a pretty good
head start, the engine part of it is pretty solid, and just working on
filling out the game part of things.